:recycle: Refactor and rename ParagraphBuilder instantiating from TextContent
