+++ render-wasm/src/wasm/text.rs
@@ -1,6 +1,6 @@
 use crate::mem;
 use crate::shapes::{GrowType, RawTextData, Type};
-use crate::textlayout::{auto_height, build_paragraphs_with_width};
+use crate::textlayout::{auto_height, build_paragraphs_with_width, paragraph_builders_from_text};
 use crate::{with_current_shape, with_current_shape_mut, STATE};
 
 #[no_mangle]
@@ -44,7 +44,7 @@ pub extern "C" fn get_text_dimensions() -> *mut u8 {
         if let Type::Text(content) = &shape.shape_type {
             // 1. Reset Paragraphs
             let paragraph_width = content.width();
-            let mut paragraphs = content.to_paragraphs(None, None, None);
+            let mut paragraphs = paragraph_builders_from_text(content, None, None, None);
             let built_paragraphs = build_paragraphs_with_width(&mut paragraphs, paragraph_width);
 
             // 2. Max Width Calculation
@@ -56,12 +56,14 @@ pub extern "C" fn get_text_dimensions() -> *mut u8 {
             // 3. Width and Height Calculation
             match content.grow_type() {
                 GrowType::AutoHeight => {
-                    let mut paragraph_height = content.to_paragraphs(None, None, None);
+                    let mut paragraph_height =
+                        paragraph_builders_from_text(content, None, None, None);
                     height = auto_height(&mut paragraph_height, paragraph_width).ceil();
                 }
                 GrowType::AutoWidth => {
                     width = paragraph_width;
-                    let mut paragraph_height = content.to_paragraphs(None, None, None);
+                    let mut paragraph_height =
+                        paragraph_builders_from_text(content, None, None, None);
                     height = auto_height(&mut paragraph_height, paragraph_width).ceil();
                 }
                 GrowType::Fixed => {}
